[ISSUE-48] Migrate javascript files to Webpack
