+++ app/javascript/packs/application.js
@@ -14,5 +14,9 @@
 //
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
+import Rails from 'rails-ujs';
+import './topbar';
+
+Rails.start();
 
 console.log('Hello World from Webpacker')
